refactor: update account store import path for consistency across routes
